+++ lib/scr/trim_editor_painter.dart
@@ -137,10 +137,8 @@ class TrimEditorPainter extends CustomPainter {
     }
 
     canvas.drawRect(rect, borderPaint);
-    canvas.drawCircle(
-        startPos + Offset(0, endPos.dy / 2), circleSize, circlePaint);
-    canvas.drawCircle(
-        endPos + Offset(0, -endPos.dy / 2), circleSize, circlePaint);
+    canvas.drawCircle(startPos + Offset(0, endPos.dy / 2), circleSize, circlePaint);
+    canvas.drawCircle(endPos + Offset(0, -endPos.dy / 2), circleSize, circlePaint);
   }
 
   @override