+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.1 - beta
+
+* Improve the file structure of the package
+* Now, you just have to import one file for using the package
+
 ## 0.3.0 - beta
 
 * Update the plugin versions